move `migrate-optimize-modifier` to core
